requirements.txt
Flask
gunicorn
Werkzeug
# Librería para autenticación biométrica (Face ID, huella, etc.)
webauthn>=1.9.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.25.0